#430: widen the cascade AMOUNT plane to 64-bit (width is per unit, like scale)
All cascade values being 32-bit int was a mistake, and the tree was already
saying so in two places.

`long` IS NOT WIDER HERE. On VC7.1/x86 `long` is 32 bits -- the tree typedefs
int64_t as `long long` for exactly that reason -- so the calc surface was
written in `long` accumulators throughout and was int32 the whole time. It read
as deliberate headroom and provided none, which is how the ceiling stayed
invisible.

AND A PARALLEL 64-BIT TWIN ALREADY EXISTED: getModifiedIntValue64, intSqrt64,
intPow64, applySMRank64 sit beside their 32-bit originals, and the money plane
(calculatePreInflatedCosts / getFinalExpense / getInflationCost /
calcCorporateMaintenance / getHurryGold / getRealPopulation) is already int64_t.
That is the fudge-factor signature one level up: a duplicated surface exists
because two WIDTHS meet, exactly as a magic constant exists because two SCALES
meet. There was a narrowing seam between the cascade and everything it feeds.

WIDTH IS PER UNIT, the same line scale is decided on: an AMOUNT accumulates
across sources and scopes at x100 and carries int64_t (the package's flat
dictionary, the receiver sums, every calc accumulator/parameter/return); a
PERCENT is a small whole number by ruling and stays int. Storage is per-unit;
the calc surface is 64-bit throughout so no intermediate can overflow.

⚑ The truncation was not only in the slot type: gt_landValue carried an
explicit `(int)` cast landing every accumulated amount INTO storage, so widening
the slot alone would have changed nothing.

SAVE COST: ZERO. The cascade serializes nothing, so this whole change needs no
migration. Width only becomes a save question for a serialized amount field, and
that path is now spec'd (save.md 8: read-old / write-new, built from the
existing soft-add + soft-remove primitives) for the fields that turn out to need
it.

Spec gaps closed in the same change: fixed-point-and-scales.md 1b (width, and
the `long` trap that hid it), save.md 8 (the widening migration).

Assert build: no error names any widened symbol.

+++ Sources/Cascade/CvCascadePackage.h
@@ -42,9 +42,13 @@ struct CvCascadePackage
 	CvCascScope scope;                     // which layout this package lives on (set at bind)
 	int identityFirst;                     // the SERVED identity, interpreted per scope (CvCascadeSlotValues.h)
 	int identitySecond;                    // its second axis: city id / team id / plot y (-1 = none)
-	mutable std::vector<int> flat;      // dictionary 1: the channel-indexed x100 flat sums (local slots)
-	mutable std::vector<int> percent;   // dictionary 2: the channel-indexed x100 percent sums (local slots)
-	mutable std::vector<int> sum;       // the receiver slots: the realized x100 totals this scope consumes
+	// ⛔ WIDTH IS PER UNIT, exactly as SCALE is ([DEC-fixedpoint-x100]). An AMOUNT accumulates -- across
+	// sources, across scopes, at ×100 -- so it carries 64 bits; a PERCENT is a small whole number by ruling
+	// (no decimals, hence no ×100), so it has nothing to accumulate into and stays 32. Widening only the
+	// amount side is what lets the combine keep its shape while the overflow ceiling disappears.
+	mutable std::vector<int64_t> flat;   // dictionary 1: the channel-indexed x100 flat sums (local slots)
+	mutable std::vector<int> percent;    // dictionary 2: the channel-indexed percent sums (local slots)
+	mutable std::vector<int64_t> sum;    // the receiver slots: the realized x100 totals this scope consumes
 	CvDerivedCacheSet<TOwner, int64_t> set;   // THE dirty protocol -- the one component, 64-bit mask form
 
 	CvCascadePackage() : scope(CASC_SCOPE_CITY), identityFirst(-1), identitySecond(-1) {}
@@ -71,7 +75,7 @@ struct CvCascadePackage
 	// ---- emit gets found ([DEC-no-self-heal]). A never-authored channel answers 0 without any storage
 	// ---- existing anywhere. ----
 
-	int readFlat(int iChannel) const
+	int64_t readFlat(int iChannel) const
 	{
 		const int iSlot = CascadeChannelRegistry::scopeSlotIndex(scope, iChannel);
 		if (iSlot < 0)
@@ -91,7 +95,7 @@ struct CvCascadePackage
 		return iSlot < (int)percent.size() ? percent[iSlot] : 0;
 	}
 
-	int readSum(int iChannel) const
+	int64_t readSum(int iChannel) const
 	{
 		const int iReceiver = CascadeChannelRegistry::scopeReceiverIndex(scope, iChannel);
 		if (iReceiver < 0)
@@ -132,7 +136,7 @@ struct CvCascadePackage
 	// ---- untouched. It is also what makes the load drain order-free: every banked mark rebuilds its own
 	// ---- inputs first, whatever order the drain walks the owners in. ----
 
-	int sourceFlat(int iChannel) const
+	int64_t sourceFlat(int iChannel) const
 	{
 		const int iSlot = CascadeChannelRegistry::scopeSlotIndex(scope, iChannel);
 		if (iSlot < 0)
@@ -154,7 +158,7 @@ struct CvCascadePackage
 		return iSlot < (int)percent.size() ? percent[iSlot] : 0;
 	}
 
-	int sourceSum(int iChannel) const
+	int64_t sourceSum(int iChannel) const
 	{
 		const int iReceiver = CascadeChannelRegistry::scopeReceiverIndex(scope, iChannel);
 		if (iReceiver < 0)
@@ -207,9 +211,9 @@ struct CvCascadePackage
 		}
 	}
 	// Slot write access for the gather (refs into the mutable storage; game-thread only).
-	int& slotFlat(int iSlotIndex) const { return flat[iSlotIndex]; }
+	int64_t& slotFlat(int iSlotIndex) const { return flat[iSlotIndex]; }
 	int& slotPercent(int iSlotIndex) const { return percent[iSlotIndex]; }
-	int& slotSum(int iReceiverIndex) const { return sum[iReceiverIndex]; }
+	int64_t& slotSum(int iReceiverIndex) const { return sum[iReceiverIndex]; }
 
 private:
 	CvCascadePackage(const CvCascadePackage&);              // noncopyable (the set binds an owner pointer)
